+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to Rainbow-CLI will be documented in this file.
 ### [1.40.0] - 2018-04-29
 - RQRAINB-889: Add commands for retrieving developer's subscriptions
 - RQRAINB-888: Add commands for retrieving developer's billing account and payment methods
+- RQRAINB-890: Extract API metrics in CSV
 
 ### [1.39.0] - 2018-04-08
 - RQRAINB-830: Inject official appID/appSecret in Rainbow CLI